requirements.txt
Flask
Flask-CORS
openai
python-docx
fpdf